RISC-V: Only cap the upper/end usable memory window
When loading the initrd in the kexec_load flow, memory for the
segments are searched from end to start. Only the max_usable should be
capped, after a successful initrd addtion.

Currently min/max usable is set to the same value, making it
impossible from subsequent segment allocations to success.

[ Björn: NB! This should be folded into another patch! ]

Signed-off-by: Björn Töpel <bjorn@rivosinc.com>